+++ packages/expo-brownfield/plugin/src/ios/plugins/withXcodeProjectPlugin.ts
@@ -39,6 +39,8 @@ const withXcodeProjectPlugin: ConfigPlugin<PluginConfig> = (config, pluginConfig
       'Messaging.swift',
       // State proxy
       'State.swift',
+      // State wrapper
+      'StateWrapper.swift',
       //SwiftUI brownfield entrypoint
       'ReactNativeView.swift',
       // UIKit brownfield view controller