+++ packages/expo-widgets/ios/Widgets/DynamicView.swift
@@ -81,18 +81,21 @@ public struct WidgetsDynamicView: View, ExpoSwiftUI.AnyChild {
         switch kind {
         case .widget:
           render(WidgetButtonView.self, ButtonProps.self) { buttonProps in
+            try updateChildren(buttonProps)
             buttonProps.source = source
             buttonProps.entryIndex = entryIndex
           }
         case .liveActivity:
           render(LiveActivityButtonView.self, ButtonProps.self) { buttonProps in
+            try updateChildren(buttonProps)
             buttonProps.source = source
           }
         }
       } else {
         render(ExpoUI.Button.self, ExpoUI.ButtonProps.self, updateProps: updateChildren)
       }
-
+    case "react.fragment":
+      render(FragmentView.self, FragmentProps.self, updateProps: updateChildren)
     default:
       ZStack {
         Color.red.opacity(0.5)
@@ -127,8 +130,9 @@ public struct WidgetsDynamicView: View, ExpoSwiftUI.AnyChild {
   private func updateChildren<P>(_ initialProps: P) throws
   where P: UIBaseViewProps {
     if let props = node["props"] as? [String: Any] {
-      if let children = props["children"] as? [[String: Any]] {
-        initialProps.children = children.map { WidgetsDynamicView(source: source, kind: kind, node: $0, entryIndex: entryIndex) }
+      if let children = props["children"] as? [Any] {
+        let validChildren = children.compactMap { $0 as? [String: Any] }
+        initialProps.children = validChildren.map { WidgetsDynamicView(source: source, kind: kind, node: $0, entryIndex: entryIndex) }
       } else if let child = props["children"] as? [String: Any] {
         initialProps.children = [WidgetsDynamicView(source: source, kind: kind, node: child, entryIndex: entryIndex)]
       }