+++ packages/expo-widgets/ios/Widgets/TimelineEntry.swift
@@ -3,6 +3,6 @@ import WidgetKit
 public struct WidgetsTimelineEntry: WidgetKit.TimelineEntry {
   public let date: Date
   public let source: String
-  public let node: [String: Any]?
+  public let props: [String: Any]?
   public let entryIndex: Int?
 }